Change the Landing page of user and made it more modular and reusable
