+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@react-google-maps/api": "^2.20.6",
+    "@reduxjs/toolkit": "^2.6.0",
     "@tailwindcss/vite": "^4.0.6",
     "@types/leaflet": "^1.9.16",
     "@types/react-leaflet": "^2.8.3",
